Game mechanics all complete
Added functionality where if you click a tile with no bombs around it,
then all safe tiles around it are also shown to the player. Also added
functionality where if a user clicks a number and the number is equal to
the number of flags around it, then all tiles around it (excluding the
flagged tiles) are clicked. Some documentation needed, otherwise this
project is mostly complete now